move communication language modes to choices package
